# api/openapi.yaml
openapi: 3.0.3
info:
  title: Booking Platform API
  version: 1.0.0
  description: |
    Multi-category service reservation platform.

    Four surfaces, four different middleware chains:

    | Prefix              | Requires                                       |
    |---------------------|------------------------------------------------|
    | `/api/v1/public`    | nothing (an optional bearer token personalizes) |
    | `/api/v1/me`        | a bearer token, scoped by its `sub`             |
    | `/api/v1/org`       | a bearer token **and** `x-organization-id`      |
    | `/api/v1/admin`     | a bearer token whose `platformRole` is `admin`  |

    **Conventions**

    - Money is always an integer in minor units (`priceCents`); 24500 = 245.00 ETB.
    - Times are always RFC3339 **UTC** on the wire.
    - Times of day (opening hours) are `"HH:MM"` strings in the provider's zone.
    - Errors are always `{ "message": string, "code": string }`.

    **Versioning.** Once the mobile app ships, `/api/v1` is additive-only:
    new optional fields are fine; removing, renaming, retyping a field or
    tightening validation is not.

    **Not yet implemented** (build order milestones 8-11): provider discovery
    search, media uploads, reviews, favourites, notifications and payments.
servers:
  - url: http://localhost:8081
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/public/providers/{slug}:
    get:
      tags: [public]
      summary: Get an active provider by slug
      description: Providers that are not `active` are reported as not found.
      security: []
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
          example: shine-wash
      responses:
        "200":
          description: The provider.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Provider" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/public/providers/{providerId}/services:
    get:
      tags: [public]
      summary: List a provider's active services
      security: []
      parameters:
        - $ref: "#/components/parameters/ProviderId"
      responses:
        "200":
          description: Active services.
          content:
            application/json:
              schema:
                type: object
                properties:
                  services:
                    type: array
                    items: { $ref: "#/components/schemas/Service" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /api/v1/me/providers:
    post:
      tags: [me]
      summary: Register a business
      description: |
        Creates a `pending` provider and makes the caller its owner. A platform
        admin decides whether it goes live.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateProviderRequest" }
      responses:
        "201":
          description: Registered, awaiting approval.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Provider" }
        "409":
          description: "`CONFLICT` — that slug is taken."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422": { $ref: "#/components/responses/ValidationError" }
...
  /api/v1/admin/providers:
    get:
      tags: [admin]
      summary: List providers across all tenants
      parameters:
        - name: status
          in: query
          schema: { $ref: "#/components/schemas/ProviderStatus" }
        - name: search
          in: query
          schema: { type: string }
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Providers, newest first.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Page"
                  - type: object
                    properties:
                      providers:
                        type: array
                        items: { $ref: "#/components/schemas/Provider" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [admin]
      summary: Onboard a provider directly
      description: For businesses signed up offline; it has no owner until invited.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateProviderRequest" }
      responses:
        "201":
          description: Created, pending approval.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Provider" }
  /api/v1/admin/providers/{providerId}:
    get:
      tags: [admin]
      summary: Get a provider and its members
      parameters: [{ $ref: "#/components/parameters/ProviderId" }]
      responses:
        "200":
          description: The provider and its staff.
          content:
            application/json:
              schema:
                type: object
                properties:
                  provider: { $ref: "#/components/schemas/Provider" }
                  members:
                    type: array
                    items: { $ref: "#/components/schemas/Member" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        A 15-minute access token, verified locally on every request. Refresh it
        with `POST /api/v1/auth/refresh`.
  parameters:
...
    ProviderId:
      name: providerId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    Offset:
      name: offset
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
  responses:
...
    Forbidden:
      description: "`FORBIDDEN` — authenticated, but not allowed to do this."
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    NotFound:
      description: "`NOT_FOUND`"
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    ValidationError:
      description: "`VALIDATION_ERROR`, with per-field detail where applicable."
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      required: [message, code]
      properties:
        message:
          type: string
          description: Human-readable and safe to show a user.
          example: That slot was just taken
        code:
          type: string
          description: |
            Stable machine-readable code. Clients switch on this, so it is part
            of the contract and additive-only.
          enum:
            - INVALID_CREDENTIALS
            - EMAIL_TAKEN
            - UNAUTHENTICATED
            - FORBIDDEN
            - NOT_A_MEMBER
            - TOKEN_EXPIRED
            - INVALID_TOKEN
            - VALIDATION_ERROR
            - ORG_REQUIRED
            - NOT_FOUND
            - SLOT_TAKEN
            - OUTSIDE_HOURS
            - INVALID_TRANSITION
            - SERVICE_INACTIVE
            - PROVIDER_INACTIVE
            - CONFLICT
            - INTERNAL_ERROR
          example: SLOT_TAKEN
        details:
          type: object
          additionalProperties: { type: string }
          description: |
            Per-field problems, keyed by attribute. Present on validation
            failures only; a client that reads just `message` and `code` is
            unaffected by it.
...
    Page:
      type: object
      properties:
        total: { type: integer, format: int64 }
        limit: { type: integer }
        offset: { type: integer }
...
    MemberRole:
      type: string
      enum: [owner, admin, staff]
...
    Member:
      type: object
      properties:
        providerId: { type: string, format: uuid }
        userId: { type: string, format: uuid }
        role: { $ref: "#/components/schemas/MemberRole" }
        name: { type: string }
        email: { type: string, format: email }
        phone: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
...
    Location:
      type: object
      required: [lat, lng]
      properties:
        lat: { type: number, format: double, minimum: -90, maximum: 90 }
        lng: { type: number, format: double, minimum: -180, maximum: 180 }
    ProviderStatus:
      type: string
      enum: [pending, active, suspended]
    BookingMode:
      type: string
      enum: [request, instant]
      description: |
        `instant` skips `pending` and confirms at creation; `request` waits for
        the provider to approve.
    Provider:
      type: object
      required: [id, slug, name, status, timezone, bookingMode]
      properties:
        id: { type: string, format: uuid }
        slug: { type: string }
        name: { type: string }
        phone: { type: string, nullable: true }
        email: { type: string, nullable: true }
        description: { type: string, nullable: true }
        city: { type: string, nullable: true }
        address: { type: string, nullable: true }
        location: { $ref: "#/components/schemas/Location" }
        timezone:
          type: string
          description: IANA zone; all slot maths happens here.
          example: Africa/Addis_Ababa
        logoUrl: { type: string, nullable: true }
        licenseNumber: { type: string, nullable: true }
        status: { $ref: "#/components/schemas/ProviderStatus" }
        ratingAvg: { type: number, format: double }
        ratingCount: { type: integer }
        bookingMode: { $ref: "#/components/schemas/BookingMode" }
        minLeadMinutes:
          type: integer
          description: Slots sooner than this from now are not offered.
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CreateProviderRequest:
      type: object
      required: [name]
      properties:
        name: { type: string }
        slug: { type: string, description: Derived from the name if omitted. }
        phone: { type: string }
        email: { type: string, format: email }
        description: { type: string }
        city: { type: string }
        address: { type: string }
        location: { $ref: "#/components/schemas/Location" }
        timezone: { type: string, default: Africa/Addis_Ababa }
        licenseNumber: { type: string }
        bookingMode: { $ref: "#/components/schemas/BookingMode" }
        minLeadMinutes: { type: integer, minimum: 0 }
        categoryIds:
          type: array
          items: { type: string, format: uuid }
...
    Service:
      type: object
      required: [id, providerId, categoryId, name, priceCents, currency, durationMinutes, isActive]
      properties:
        id: { type: string, format: uuid }
        providerId: { type: string, format: uuid }
        categoryId: { type: string, format: uuid }
        name: { type: string }
        description: { type: string, nullable: true }
        priceCents:
          type: integer
          description: Minor units. 24500 = 245.00 ETB.
        currency: { type: string, default: ETB }
        durationMinutes:
          type: integer
          description: The scheduling engine's only real input.
        bufferMinutes:
          type: integer
          description: Added after the service; occupies the resource but is not priced.
        attributes:
          type: object
          additionalProperties: true
          description: Validated against the category's `service` definitions.
        imageUrl: { type: string, nullable: true }
        isActive: { type: boolean }
        optionGroups:
          type: array
          description: Present on the detail endpoints only.
          items: { $ref: "#/components/schemas/OptionGroup" }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...